+++ modules/components/widgets/antd/value/Select.jsx
@@ -1,6 +1,7 @@
 import React, { PureComponent } from "react";
 import PropTypes from "prop-types";
-import {useOnPropsChanged, mapListValues, calcTextWidth, SELECT_WIDTH_OFFSET_RIGHT} from "../../../../utils/stuff";
+import {mapListValues, calcTextWidth, SELECT_WIDTH_OFFSET_RIGHT} from "../../../../utils/stuff";
+import {useOnPropsChanged} from "../../../../utils/reactUtils";
 import { Select } from "antd";
 const Option = Select.Option;
